refactor: Infrastructure インターフェースを各実装フォルダに移動
- IOutputDirectoryProvider.cs → Infrastructure/FileSystem/
- IGitHubEnvironmentUpdater.cs → Infrastructure/GitHub/
- IFileDownloader.cs           → Infrastructure/Http/
- INotionClientWrapper.cs      → Infrastructure/Notion/
